feat: ini handle create swap
