+++ packages/hooks/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@lightdotso/client": "workspace:*",
     "@lightdotso/data": "workspace:*",
+    "@lightdotso/nuqs": "workspace:*",
     "@lightdotso/params": "workspace:*",
     "@lightdotso/query": "workspace:*",
     "@lightdotso/schemas": "workspace:*",
